Create BMPersist_File_script.ps1
This script is used to read the current file count and size of BMPersist folder in the primary and redundant batch executive(s). If the folder size and file count match in both servers, than a 'True' bit is written to a .csv file, along with PI tag name and time stamp.